+++ test/dunolint-lib/test__dune_workspace__dune_lang_version.ml
@@ -143,7 +143,9 @@ let%expect_test "Predicate.t_of_sexp" =
   [%expect
     {|
     (Of_sexp_error
-     "dune_lang_version.t_of_sexp: polymorphic variant tag takes an argument"
+     (Dunolint.Sexp_helpers.Error_context.E
+      ("The construct [=] expects one or more arguments."
+       (suggestion "Replace by: (= ARG)")))
      (invalid_sexp =))
     |}];
   test "(= 3.17 extra)";
@@ -156,7 +158,10 @@ let%expect_test "Predicate.t_of_sexp" =
   test "(unknown 3.17)";
   [%expect
     {|
-    (Of_sexp_error "dune_lang_version.t_of_sexp: no matching variant found"
+    (Of_sexp_error
+     (Dunolint.Sexp_helpers.Error_context.E
+      ("Unknown construct [unknown]."
+       (did_you_mean? ((var unknown) (candidates = > >= < <= !=)))))
      (invalid_sexp (unknown 3.17)))
     |}];
   ()